IO system ready
Images can be converted into nodes, and nodes can be outputted as printed strings. Start and Target positions are marked. The A-star function is ready for testing. 2D and 3D hardcoding was removed and replaced with any-dimension general functions. Location information will be stored as lists for the time being.